storage/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    API_ID = int(os.getenv("API_ID", 0))
    API_HASH = os.getenv("API_HASH", "")
    SESSION_NAME = os.getenv("SESSION_NAME", "session")
    DOWNLOAD_PATH = os.getenv("DOWNLOAD_PATH", "./downloads")

    # Tạo thư mục download nếu chưa có
    if not os.path.exists(DOWNLOAD_PATH):
        os.makedirs(DOWNLOAD_PATH)